Added test for baggage allowance

+++ src/main/java/Plane.java
@@ -13,7 +13,7 @@ public class Plane {
         return type.getCapacity();
     }
 
-    public int getTotalWeight() {
+    public double getTotalWeight() {
         return type.getTotalWeight();
     }
 }
